Remove unnecessary git subcommand
It is confusing to use `git remote show` to list the remotes.

The git command `remote`, when used with no arguments, shows a list of
existing remotes:

https://git-scm.com/docs/git-remote#_commands

The `show` subcommand is not necessary and might be confusing to use it
as it is used to give information about a given remote, i.e.,  `git
remote show heroku`:

https://git-scm.com/docs/git-remote#Documentation/git-remote.txt-emshowem

As all the information that is needed is the list of remotes to check if
`heroku` is there, the command should be `git remote`.