Removed method reference to previous lesson
Referring to numbered list item 3 under heading Strings which states as follows:

3. Vocabulary time: a __method__ is a bit of functionality that is built into the language or into specific data types.  In the previous lesson you learned a few methods that can be used on strings such as `indexOf` and `search`.  An exhaustive list of methods that can be used on strings can be found [here](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String).

The second sentence is incorrect as it states that methods were discussed in the previous lesson, however this is not true. Methods are only introduced in the numbered list item 2 which precedes this item. I believe it is most likely that there was a hyperlink in the previous lesson where methods are touched on, however this was removed at a later stage.

This sentence could create confusion for the student who may end up revising all prior work to see where they had missed the information on methods, only to discover that there was no information provided.

One can change the sentence to read that methods were introduced in the preceding numbered point, however this could lead to the same issue if the preceding numbered list item is altered and the method information removed.

It makes more sense to me to remove the sentence that refers to where the individual had found the information on methods: 
"In the previous lesson you learned a few methods that can be used on strings such as `indexOf` and `search`. "

The paragraph would then read as follows, as reflected by the Push request:
"3. Vocabulary time: a __method__ is a bit of functionality that is built into the language or into specific data types. An exhaustive list of methods that can be used on strings can be found [here](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String).