Update DOM-manipulation.md
There was a dead link in the section, specifically the one relating to arrow functions in the event listener section. This one : http://javascript.info/function-expressions-arrows#arrow-functions

I found this one instead. http://javascript.info/arrow-functions-basics

I am not sure if this is similar to the other one. Since it is down. I leave it to you professionals to see if this is correct.