Replaced submission statement and added link to repo
The submission statement on this page doesn't have a link to the repo page that must be altered, nor does it have a link to the contributing guide. 

Switched out the submission statement with the one found on other pages that contains both a link to the repo page that must be edited, as well as the hyperlink to the Contributing guide.