Change out Klavarog for more user friendly app
I could not for the life of me get Klavarog to work until I registered an email, clicked the link and then finally did I have an option to load in a premade dictionary. Keybr.com seems like a great replacement for anyone who wants to improve their skills. Straight forward instructions, registration is easy and not entirely necessary.