add link (#16651)
got kind of confused myself by trying to nest params in a hash using the older form_tag, as the current guides don't really cover it and neither does the "form tag api documentation" in the link under #4.  Added a link to the older rails guide docs that was helpful.