Update problem_solving.md
Fixed more errors. Glad this is helpful. I will leave the PR template in the future but for this one I wanted to double check one change. For the "Understanding The Problem" section I believe the first line meant to say "Write a program" but I could be incorrect if the original "Write a problem" has a particular meaning here. Just wanted to clarify that one.

Thanks,
Daniel Murphy