Update problem_solving.md
Fixed this page to more closely match the changes to the Ruby Problem Solving page. We will need to see if the new tilde formatting fixes the line spacing problem in the pseudo code blocks (issue #18528). Note that some changes here were not exactly errors but just done to match the Ruby page and maintain consistency. I will go back to the Ruby page now and fix a couple other errors I had missed the first go around and correct the quote attribution to V. Anton Spraul who was listed as V. Anton Paul on the Ruby page.