Fix `Postman` link
The previous link had either an incorrect title or incorrect URL as they
referred to different tools. It seems that Chrome's rest client
extension has been deprecated and they recommend to install desktop
client [1]. As this is the case, include Postman as recommended tool
(which the link directed to before as well).

[1] https://chrome.google.com/webstore/detail/advanced-rest-client/hgmloofddffdnphfgcellkdfbfbjeloo