Update state-and-props.md
Updated Line 61 for improved clarity about the example below it. Changed 'this works with functions' to 'props work with functions'. On my initial read I thought it was referring to function components vs class components. It wasn't until I looked at the example that I realized the sentence was referring to passing functions as props.