Update state-and-props.md
Line 61. Further update to that paragraph for clarity and grammar. Changed 'it works the same' to 'they work the same way' to remove repetitive 'it' and complete the sentence with 'way'. It feels like a more complete thought this way. Sorry for not catching it in yesterday's edit. Thanks for what you do!