Add Figma as a resource for HTML CSS course
Figma is brought up multiple times a week in response to people asking what a good alternative to laying out a site would be. Keep it DRY.
Could this sentence describing Figma give any more context to help the reader understand what it's for? I tried to stay with the same single sentence format as the other links.