Stressed importance of commiting early and often
A lot of students don't quite grasp the importance of committing early and often.
Briggs gave me the go-ahead in Discord to make a PR somewhere about stressing its importance.
I wasn't sure if this was the best place to put it. However, I think it seems the most fitting as it adds on to the previous paragraph nicely.