Restore original JS Prototype article link
I followed the rabbit hole with this website in the closed PRs. It seems like the site is doing OK.

https://web.archive.org/web/20200513181548/https://javascriptissexy.com/javascript-prototype-in-plain-detailed-language/ is the archive link.

If it does go down again, could make a rule that this site just stays as an archive.

Reasons to change it back? It takes forever to load, the site doesn't get the hits it deserves (when it's online).