Update clean_code.md
The examples given contradict the advice for naming variables, which states that they must begin with a noun. "My" in "myName" is an adjective/determiner and "selected" is a past participle (i.e., a kind of adjective). Variable names often start with adjectives and constitute noun phrases.

Along these lines, the advice given here is also contradicted by the suggested name of one of functions in the Rock Paper Scissors project of the previous lesson, viz computerPlay(), which begins with a noun. [g]ame() could also be construed as a noun.