Update nested_collections.md
The language about trying to index a nonexistent nested element vs accessing a nonexistent index or key of an existing nested element was not very clear or specific. I believe this edit more succinctly clears up the intention of the section! 

Change applied to both the array and hash discussions surround the #dig method.