Reorganize installation overview os options into dropdowns
Some students are seeming to not realize that you can do odin even if
you are currently using windows. In an effort to clarify that the
different OS options presented to students in this lesson are now
reorganized into dropdowns, with a windows heading on the dropdown that
is directed toward students currently using windows.