Update conclusion.md
Made some grammar changes and also fleshed out some of the acronyms. This gives the reader a little more to go off. As a beginner and a learner, I don't like encountering specialist jargon or acronyms with zero context or explanation. I think that a course for beginners (or any course, really) should assume that the learner is not versed in computerese.