Update How it works
Added a paragraph to the How It Works section. Since a lot of students seem to be voicing the same concern over forgetting material after having finished FCC, and not feeling ready for the Google Project, it seems appropriate to explain that memorization is not an expectation, and that programming is very different from the educational model most people are used to (school, studying, notes, tests, failure/passing).