Update es6-modules.md
I made a few grammar edits and word changes. I think "compile" is best avoided when explaining webpack, because of its other, technical meaning of converting code to binary. So I made some substitutions.

It seemed odd to me that named exports were exemplified, while there was nothing on default ones, so I added a brief demo and explanation for the latter.