Rename Developer Tools 2 - Foundations
After the arrival of new lessons old Developer Tools lesson is going to be deleted. So Developer Tools 2 lesson name in JS Basics no longer makes sense. So the Pull Request created to change the lesson name in the site repo. This Pull Request changes the `md` file name to the new name.