Using Heroku with the Apple M1
This request is to help users understand how to install and use Heroku on Macs that have the Apple M1 chip. I'm submitting it because as I was following the instructions in the course, I realized that I was unable to install Heroku and found the article that I linked really helpful to enable me to keep moving through the lesson. 

I also realize there is value in making people solve the problem themselves but this change may allow them to keep moving in the curriculum.

Thanks for the consideration and for the curriculum.