Fix wrong conjunction: replace 'and' with 'so'.
In this sentence:
"There’s probably a good reason why they need more information, and trust their judgement and experience when they ask."
"...and trust" implies that the object of the sentence is still "they", so it reads a bit confusingly. Using "so" makes it clear that the second part of this sentence is a command directed at the reader.