Added statement in head for the removed statement.
Added the correct statement for the previously removed statement. I was thinking to give the quick idea to the readers that if we shouldn't add the displaying elements in <head> then we have <body> element where we can use them. But later I thought, let them keep engaging in reading to figure out the answer when they'll read the information under Body Element.