Added changes about images in links-and-images.md
We just display an existing image using the <img> tag and do not create any new image using it. The word "create" gives the wrong sense that we are generating an image. That's why I think the "display" word can properly convey the use of <img> tag i.e. just to display an existing image.