Basic Data Types: Fix broken Assignemnt Link
The second link in the assignment sections is a bad link. This is the last good archive of that page available from archive.org. It does appear the URL was taken over some time recently as the whois data has changed as of 04/07/22 however archive.org was getting 301 errors for this page as recently as March 2020.