Change a link of the "Knowledge Check" section
The last question of the "Knowledge Check" section asks: How do you attach event listeners to elements in React components? The link sends you to a page of the React documentation that shows you all sorts of supported events but there is no information of how to attach those events to elements. Instead, the new link sends the user to a section of the documentation that shows you exactly how to attach and handle events in React.