Additional Resources sections: Remove leading whitespace before paragraphs
This commit removes the leading whitespace before each paragraph/list in
the page section "Additional Resources" for consistency with the the rest
of the text in each lesson.

This change does not affect the live html result of the markdown source,
but simply improves the consistency of the source file for future
contributors.

See discussion: https://github.com/TheOdinProject/curriculum/issues/24332#issuecomment-1176079916