Update headings
There should be white space around all headings and sub-headings.

The headings and subheadings should follow Wikipedia Style Title Case.
Along with this, I removed puncuation since that is not allowed per
Wikipedia, Microsoft, and other projects. This does remove some of the
charm or character of TOP like the exclamation mark at the end of one
of the headings.