Git Basics: fix indentation to properly render list
In the Git Basics lesson's assignment, item 4 under "Use the Git Workflow" mistakenly showed the markdown to add a screenshot instead of displaying the screenshot itself. This was because the line before the image wasn't indented to the same level as the rest of the items in the assignment list. As a result, the image markdown code was seen as indented and treated as a code block instead of as a markdown element. This commit fixes the indentation which allows the markdown to properly render.

Additionally, in the same assignment, some of the list items under "Create the Repository" weren't showing the correct item numbers. Since the indentation was off, each item was being seen as a new list and the auto-numbering would restart at 1.