Node.js Learn Page Link Paths were outdated
All of the links to the Node.js Learn pages were outdated. The new URL path requires a language specification between "nodejs.dev" and "/learn". This outdated path format leads to a "page not found" error. My proposed change involves adding "/en/" in the URL path in order to link to the proper page.