Introduction to Grid: Added Video to Sup. Material
1. Fantastic visual representation of strengths of CSS Grid vs Flexbox
I've added the 'Flexbox vs. CSS Grid — Which is Better?' video from LayoutLand on Youtube because it does great job of showing a visual representation of what Flexbox and Grid do best respectively. Notably, the section from 2:38 to 3:39 shows a few different layout templates with gridlines which make it intuitively clear not just what works best in each case, but also why. 

The image in the CSS Tricks link with the label "Flexbox on top, Grid on bottom" shows the different ways that Flexbox and Grid would effect the same grid, (perhaps, one could say in a desirable fashion vs an undesirable one depending on the circumstances), but the advantage of the video is that it shows the desired outcome, and shows how CSS Grid or Flexbox would naturally make the task easier, more difficult, or even impossible to accomplish.

2. Increase diversity of Supplemental Resources 
Finally, LayoutLand provides other high quality content on layouts with CSS Grid and Flexbox and it would be advantageous to expose students to a more diverse pool of high quality content outside of the usual CSSTricks and Kevin Powell videos, for example. I mention this simply because as far as I can tell this would be the first LayoutLand resource to be used in the curriculum, and it's clear that this video and possibly other has a potential to offer value or fill in gaps where content from other creators might not, or at least expose to the students an additional teacher that they can choose from to dive into further videos past the current offering.

In summary, the video I've included provides a good visual primer to be able to better understand the strengths of Flexbox and CSS Grid going into the CSS Tricks article, and in general. It would also be good to include a video from another instructor of high quality CSS/Flexbox content to increase the diversity of the program material.