Block and inline lesson: Improve grammar
The assignment 4 points to the CSS repo and instructs to do the two exercises. At first I was confused where the exercises were located, and only after entering the different directories i figured that the exercise files were in the `margin-and-padding` directory (duh). I know it's silly and obvious in retrospect, however, if it happened to me, I'm sure it will happen to others as well.  I propose to change the wording to point people out to the right directory, just like it was done in CSS Foundations lesson assignment 1.