Update css-foundations.md specificity example
Update css-foundations.md to include a clearer example on specificity. The first group of examples on specificity have three examples with div classes and ids of "main", "list" and "subsection". The first two examples both have matching HTML and CSS. However, the third example has CSS only that does not work with the previous HTML snippet. An HTML snippet has been added and the CSS reworked very slightly for the third example, to remove ambiguity whilst preserving the intent of the example.