Update a_deeper_look_at_git.md
there is a pretty large amount of information in this page and some of it kind of flew by my head. I felt like I really needed to use the additional resources in order to get a good grasp on intermediate Git and the two page by git-scm provided the most benefit to me by the time I finished. the diagrams are easy to understand and revisiting the information really helped it set for me. I believe those two, the one on rebasing and the one on reset should be included in the main assignment section, otherwise people may leave without fully understanding the material.