Improve consistency and clarity in examples
By being consistent and explicitly labeling examples as Wrong/Right we
avoid confusing new contributors.

The use of "Title case" should be avoided when describing the suggested
case (Sentece case) for headings because it is a term that defines a
specific form of capitalization ("This Is Title Case") that goes against
that same suggestion.