Links and Images lesson: Add opening links in a new tab section (#25747)
* Added the Opening links in new tab section

* Added to Knowledge check and Additional resources for opening links in new tabs

Added a new knowledge check question to emphasize remembering rel atttribute value for security reasons. Had to make a span id link to the relating portion of the webpage. Added a link to an opinion page of when to use the target=_blank field in additional resources.

* Split some of the paragraphs in the Opening links in a new tab section

* Fixed typo

Containing spelling typo in the opening links in new tab section for the links and images lesson