Time Complexity lesson: Improve text in Quadratic/Cubic sections. (#25611)
* Clarify Quadratic, Cubic Math in 'Time Complexity'

The previous text in the Quadratic and Cubic sections of this lesson
confused the total number of sub-steps with the number of extra sub-steps,
referring ambiguously to either the sub-steps when looping over a data
structure or the number of loops.

* Reword the Quadratic/Cubic paragraphs to be more..

friendly for those not from a mathematical background. Numerical
examples written to communicate the speed of growth in the total
number of sub-steps when the size of an array increases.

* Correct a typo on 10 cubed = 1000 sub-steps

* Remove "Therefore, when....n squared sub-steps" sentences from Quadratic section

* Change the Quadratic and Cubic explanations in ...

the Javascript Time Complexity lesson to match the new version of each
in the corresponding lesson of the Ruby path.