Markdownlint: Add custom rule  "Blanks around multiline HTML tags" (#27674)
* Add test md file for rule

* Add custom TOP005 rule

* Include code block delimiter and ignore-comment tests

* Allow code block delimiters and ignore-comments

Original function would flag markdownlint-disable-* comments which would
require being followed by the line to disable.
It would also flag HTML tags directly following or preceding triple
backticks. These are now allowed so that code blocks can start and end
more cleanly.

* Add TOP005 documentation file

Added TOP005 rule to linter config file.

* Disable mdlint for documentation first lines

These files' first lines need to be level 1 headings, not level 3, as
they are not files to be rendered on the TOP website.

* Adjust verbiage for rationale

* Fix TOP005_test.md heading level error

* Account for first/last line

Technically HTML tags on the first or last line of a file would trigger
other rules to error, but adding `?? ""` for the sake of completeness.

Single quotes have been switched to double quotes to be consistent with
other JS files in the repo.

* Update TOP005_test.md

Demonstrates how a lack of blank lines interferes with unrelated lint
checks.

* Add actual HTML code example to test file

* Prevent flagging rule for html/jsx code block examples

These cases have a legitimate need to "break the rule".
We only want to flag for HTML tags we use for the actual lesson markup,
or md code block examples of such.

.trim added so that indentation doesn't matter.

* Update TOP005 doc file

Now includes examples of ignored code blocks (html/jsx type).

* Amend TOP005_test.md to only demonstrate the TOP005 rule

Removed unrelated errors to make it easier to demonstrate TOP005.
The behaviour with unrelated errors are now explained in text.
Fixed opening introduction line as it incorrectly stated there would not
be any errors flagged.

* Improve readability in rule code

Opted to return token.map as is on line 19 instead of the same but with 1
subtracted from token.map[1].
The only difference this makes is include the closing backticks line in
the isWithinIgnoredFence range check, which will have no effect on the
outcome of the check.
The benefit is not needing to explain why we need to subtract 1 from
token.map[1] only. This allows everything to remain intuitive, and the
only deviation being the onError.lineNumber value which has a comment to
explain its 1-index requirement.

Regexr test links added.

Co-authored-by: Eric Olkowski <thatblindgeye@gmail.com>

* Remove markdownlint-disable comments from doc files

These should already be ignored by the lint actions

* Improve clarity of onError.detail message

Includes visual formatting improvements.

---------

Co-authored-by: Eric Olkowski <thatblindgeye@gmail.com>