Setting Up a React Environment: Remove "Keeping it clean" section (#28306)
`eslint-config-prettier` is only needed if using an ESLint ruleset that turns on clashing style rules.
These are deprecated and not included in the default ruleset, so by default, will not be needed.
It would only be needed if someone specifically uses an external plugin that turns on those rules,
like `airbnb`.

This is covered in the JS Linting lesson already, and feels out of place to mention here, especially if going into the airbnb/default ruleset caveats. People will continue to use what they've already been using, and can always revisit the Linting lesson as necessary.